+++ uC_api/uC.py
@@ -61,11 +61,11 @@ class uC_api:
         """
         self.__experiment_state = []
         self.__experiment_state_timestamp = []
-        self.__connection = serial.Serial(serial_port_path,115200, timeout= None) #its USB so the speed setting gets ignored and it runes at max speed
+        connection = None
         self.__read_buffer = Queue()
         self.__write_buffer_timed = Queue()
         self.__write_buffer = Queue()
-        self.__communication_thread = threading.Thread(target=self.__thread_function)
+        self.__communication_thread = threading.Thread(target=self.__thread_function, args=(serial_port_path,))
         self.__free_input_queue_spots_on_uc = -1
         self.__last_timed_packet = 0
         self.__api_level = api_level
@@ -230,18 +230,18 @@ class uC_api:
         self.__experiment_state.append(-1)
         self.__experiment_state_timestamp.append(-1)
 
-    def __check_first_connection(self):
+    def __check_first_connection(self,connection):
         """__check_first_connection checks if the uC is responding and prints the firmware version
         if the firmware version does not match the API version it will print a warning
         """
         logging.info("send: opening connection - aligning commuication")
         # write 9 bytes to the uC to align the communication
-        self.__connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
+        connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
         # wait for 10 seconds for the uC to align
         for i in range(40):
             # check if the uC has send a packet
-            if self.__connection.in_waiting >= 9:
-                byte_packet = self.__connection.read(size = 9)
+            if connection.in_waiting >= 9:
+                byte_packet = connection.read(size = 9)
                 # do reading alingnment
                 byte_packet = byte_packet.lstrip(b'\xff')
                 sleep(0.05)
@@ -250,14 +250,14 @@ class uC_api:
                     if len(byte_packet) == 0:
                         # no packet (was only alignment bytes), jump to next iteration
                         continue
-                    elif self.__connection.in_waiting < 9-len(byte_packet):
+                    elif connection.in_waiting < 9-len(byte_packet):
                         # something went wrong, try again
                         logging.error("partial packet received, but not enough bytes send by uC, trying to recover by realigning")
-                        self.__connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
+                        connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
                         continue
                     # complete partial packet
                     else:
-                        byte_packet = bytearray(byte_packet).extend(bytearray(self.__connection.read(size = 9-len(byte_packet))))
+                        byte_packet = bytearray(byte_packet).extend(bytearray(connection.read(size = 9-len(byte_packet))))
                 # convert the byte packet to a packet object
                 read_packet = Packet.from_bytearray(byte_packet)
                 # check if the packet is the expected Success packet
@@ -268,25 +268,27 @@ class uC_api:
                         logging.warning("uC firmware version does not match the API version: \nfirmware version: "+str(read_packet.original_header())+"."+str(read_packet.original_sub_header())+"."+str(read_packet.value())+" \nAPI version: "+str(int(FIRMWARE_VERSION.FIRMWARE_VERSION_MAJOR))+"."+str(int(FIRMWARE_VERSION.FIRMWARE_VERSION_MINOR))+"."+str(int(FIRMWARE_VERSION.FIRMWARE_VERSION_PATCH)))
                     # connection is established
                     connection = True
+                    connection_state = True
                     return True
                 else:
                     logging.warning("unknown packet received, while connecting to uC for the first time: "+str(read_packet))
             sleep(0.25)
         # connection failed after 40 tries/10sec
-        if connection == False:
+        if connection_state == False:
             logging.error("uC is not responding for 10 sec, wrong port?, no permission?")
             return False
 
 
-    def __thread_function(self):
+    def __thread_function(self,serial_port_path):
         """__thread_function internal function managing the actual async communication with the uC in the background
         """
         idle_write_pc = False
         idle_write_uc = 0
         idle_read = False
+        connection = serial.Serial(serial_port_path,115200, timeout= None, write_timeout=0) #its USB so the speed setting gets ignored and it runes at max speed
         # init communication by forcing the uC to align
-        if not self.__check_first_connection():
-            self.__connection.close()
+        if not self.__check_first_connection(connection):
+            connection.close()
             return
 
         # start communication
@@ -300,11 +302,11 @@ class uC_api:
                     data_packet = self.__write_buffer.get()
                     # check and close the connection if requested by API
                     if data_packet.header() == Data32bitHeader.UC_CLOSE_CONNECTION:
-                        self.__connection.write(Data32bitPacket(Data32bitHeader.IN_RESET).to_bytearray())
-                        self.__connection.close()
+                        connection.write(Data32bitPacket(Data32bitHeader.IN_RESET).to_bytearray())
+                        connection.close()
                         return
                     # else send the packet
-                    self.__connection.write(data_packet.to_bytearray())
+                    connection.write(data_packet.to_bytearray())
                     logging.debug("send instant: "+str(data_packet))
                     self.__write_buffer.task_done()
                 # then write the timed packets
@@ -314,7 +316,7 @@ class uC_api:
                         # send the packet and decrease the free input queue spots reference in the API
                         data_packet = self.__write_buffer_timed.get()
                         self.__free_input_queue_spots_on_uc -= 1
-                        self.__connection.write(data_packet.to_bytearray())
+                        connection.write(data_packet.to_bytearray())
                         logging.debug("send timed: "+str(data_packet))
                         self.__write_buffer_timed.task_done()
                     else:
@@ -325,16 +327,16 @@ class uC_api:
                         idle_write_uc += 1
                         if idle_write_uc%2000 == 1:
                             packet_to_send = Data32bitPacket(Data32bitHeader.IN_FREE_INSTRUCTION_SPOTS)
-                            self.__connection.write(packet_to_send.to_bytearray())
+                            connection.write(packet_to_send.to_bytearray())
                             logging.debug("send timed: "+str(packet_to_send))
             else:
                 # set write loop slowdown condition flag
                 idle_write_pc = True
 
             # check if there is a package to read from the serial connection
-            if self.__connection.in_waiting >= 9:
+            if connection.in_waiting >= 9:
                 idle_read = False
-                byte_packet = self.__connection.read(size = 9)
+                byte_packet = connection.read(size = 9)
                 # remove alignment bytes
                 byte_packet = byte_packet.lstrip(b'\xff')
                 # check if the packet is complete
@@ -343,12 +345,12 @@ class uC_api:
                         #initiate alignment for incompleate packets
                         logging.warning("outgoing uC alignment needed, shifted by "+str(len(byte_packet))+ " bytes")
                         # partial packet not recoverable
-                        if self.__connection.in_waiting < 9-len(byte_packet):
+                        if connection.in_waiting < 9-len(byte_packet):
                             logging.error("partial packet received, but not enough bytes send by uC, trying to recover by realigning")
-                            self.__connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
+                            connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
                             continue
                         # to complete partial packet
-                        byte_packet = bytearray(byte_packet).extend(bytearray(self.__connection.read(size = 9-len(byte_packet))))
+                        byte_packet = bytearray(byte_packet).extend(bytearray(connection.read(size = 9-len(byte_packet))))
                         
                     else:
                         logging.debug("alignment sucesss - no incoming alignment error")
@@ -361,7 +363,7 @@ class uC_api:
                 except:
                     # packet was malformed, force alignment sequence
                     logging.error("packet is malformed, maybe misaligned, trying to recover by realigning")
-                    self.__connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
+                    connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
                     continue
                 # packet is complete and valid
                 logging.debug("read: "+str(read_packet))
@@ -373,7 +375,7 @@ class uC_api:
                 # catch the special case of the uC reporting an malformed packet from the API
                 elif read_packet.header() is ErrorHeader.OUT_ERROR_UNKNOWN_INSTRUCTION or read_packet.header() is ErrorHeader.OUT_ERROR_UNKNOWN_CONFIGURATION:
                     logging.error("uC is reporting that it cant understand a send packet, either API and firmware are a different version or communication is not aligned, trying to recover by realigning")
-                    self.__connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
+                    connection.write(b'\xff\xff\xff\xff\xff\xff\xff\xff\xff')
                 # normal packet, send to the read buffer for further processing by the main thread
                 else:
                     self.__read_buffer.put(read_packet)